cache finally works
