+++ src/utils/ServerContext.ts
@@ -1,0 +1,13 @@
+import { cache } from "react";
+
+export default function createServerContext<T>(defaultValue: T): [() => T, (v: T) => void] {
+  const getRef = cache(() => ({ current: defaultValue }));
+
+  const getValue = (): T => getRef().current;
+
+  const setValue = (value: T) => {
+    getRef().current = value;
+  };
+
+  return [getValue, setValue];
+};
